+++ frontend/src/components/Cart/CartItem.tsx
