+++ frontend/src/components/CategoryTabs/CategoryTabsSkeleton.tsx
